Add files via upload
The html is the Python notebook, conducting exploratory analysis on AirBnB data. The data can be accessible on Kaggle: https://www.kaggle.com/stevezhenghp/airbnb-price-prediction

The EDA analysis here is to look at the distribution of the values for each columns(number of beds, cancellation policy, geographical location, etc) by data visualization. The charts generated in this notebook are interactive graphs(histogram and scatter plots on US map), which can be zoomed in or out to get extra information. 

In addition, natural language processing(NLP) is performed to analyze the text data, amenities, looking for the amenities that are provided most and generating the word cloud visualization based on that.